Add test coverage for C++ library and application plugin
