+++ subprojects/language-native/src/integTest/groovy/org/gradle/language/swift/AbstractSwiftIntegrationTest.groovy
@@ -16,9 +16,7 @@
 
 package org.gradle.language.swift
 
-import org.gradle.internal.os.OperatingSystem
 import org.gradle.language.AbstractNativeLanguageComponentIntegrationTest
-import org.gradle.nativeplatform.platform.internal.DefaultNativePlatform
 import org.gradle.util.Matchers
 import org.gradle.util.Requires
 import org.gradle.util.TestPrecondition
@@ -50,47 +48,14 @@ abstract class AbstractSwiftIntegrationTest extends AbstractNativeLanguageCompon
         failure.assertThatCause(Matchers.containsText("Swift compiler failed while compiling swift file(s)"))
     }
 
-    def "binaries have the right platform type"() {
-        given:
-        makeSingleProject()
-        buildFile << """
-            task verifyBinariesPlatformType {
-                doLast {
-                    ${mainComponentDsl}.binaries.get().each {
-                        assert it.targetPlatform instanceof SwiftPlatform
-                        assert it.targetPlatform.operatingSystem.name == "${OperatingSystem.current().name}"
-                        assert it.targetPlatform.architecture.name == "${DefaultNativePlatform.currentArchitecture.name}"
-                    }
-                }
-            }
-        """
-
-        expect:
-        succeeds "verifyBinariesPlatformType"
-    }
-
-    def "binaries have the right tool chain type"() {
-        given:
-        makeSingleProject()
-        buildFile << """
-            task verifyBinariesToolChainType {
-                doLast {
-                    ${mainComponentDsl}.binaries.get().each {
-                        assert it.toolChain instanceof ${toolChain.implementationClass}
-                    }
-                }
-            }
-        """
-
-        expect:
-        succeeds "verifyBinariesToolChainType"
-    }
-
     protected abstract List<String> getTasksToAssembleDevelopmentBinary()
-
-    protected abstract void makeSingleProject()
 
     protected abstract String getDevelopmentBinaryCompileTask()
 
     protected abstract String getMainComponentDsl()
+
+    @Override
+    protected String getAllBinariesOfMainComponentBuildScript() {
+        return "${mainComponentDsl}.binaries.get()"
+    }
 }
